requirements.txt
about-time==3.1.1
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
async-timeout==4.0.2
autopep8==1.5.7
backcall==0.2.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.26.28
botocore==1.29.28
Brotli==1.0.9
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
cryptography==35.0.0
cssselect2==0.6.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-environ==0.8.0
django-extensions==3.1.5
django-filter==21.1
django-import-export==2.7.1
django-infinite-scroll-pagination==1.2.1
django-rest-swagger==2.2.0
django-silk==4.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.18
drf-extra-fields==3.4.0
drf-yasg==1.21.3
et-xmlfile==1.1.0
flower==1.0.0
fonttools==4.34.4
gprof2dot==2021.2.21
grapheme==0.6.0
gunicorn==20.1.0
html5lib==1.1
humanize==3.12.0
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
JSON-log-formatter==0.5.1
kafka-python==2.0.2
kombu==5.1.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
num2words==0.5.11
numpy==1.21.5
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
pydyf==0.2.0
Pygments==2.10.0
pymongo==3.12.1
pyparsing==3.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
pyphen==0.12.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.3.4
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sentry-sdk==1.4.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zopfli==0.2.1